RT-19604: CellSkinBase -fx-cell-size StyleableProperty has a wrong default value type

+++ javafx-ui-controls/src/com/sun/javafx/scene/control/skin/CellSkinBase.java
@@ -111,7 +111,7 @@ public class CellSkinBase<C extends Cell, B extends CellBehaviorBase<C>> extends
      *                                                                         *
      **************************************************************************/
 
-    static final int DEFAULT_CELL_SIZE = 24;
+    static final double DEFAULT_CELL_SIZE = 24.0;
 
      /**
       * Super-lazy instantiation pattern from Bill Pugh.
